Update example: No KeyAuthorizations on issuance  (#29)
Using KeyAuthorizations in this way is not the intention of the
specification. KeyAuthorizations can be used by the Issuer to permit
DeviceKey to sign over certain data elements. The intention of this is
that it should be used only for some dynamic elements that the Issuer
cannot provide at time of issuance, but that it trusts the Wallet to
convey. This test was previously permitting the DeviceKey to sign over
all data elements that are in the MSO - this is unnecessary and can be
viewed as untrustworthy by the Reader.